# Add faster implementation of _IonNature.ion_hash