--- a/ionhash/__init__.py
+++ b/ionhash/__init__.py
@@ -11,14 +11,9 @@
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
 
-from six import BytesIO
-
-from amazon.ion.core import ION_STREAM_END_EVENT
 from amazon.ion.simple_types import _IonNature
-from amazon.ion.simpleion import _dump, _FROM_TYPE
-from amazon.ion.writer import blocking_writer
-from amazon.ion.writer_binary import binary_writer
 
+from ionhash.fast_value_hasher import hash_value
 from ionhash.hasher import hashlib_hash_function_provider
 from ionhash.hasher import hash_writer
 from ionhash.hasher import HashEvent
@@ -54,10 +49,7 @@ def ion_hash(self, algorithm=None, hash_function_provider=None):
     else:
         hfp = hash_function_provider
 
-    hw = hash_writer(blocking_writer(binary_writer(), BytesIO()), hfp)
-    _dump(self, hw, _FROM_TYPE)
-    hw.send(ION_STREAM_END_EVENT)
-    return hw.send(HashEvent.DIGEST)
+    return hash_value(self, hfp)
 
 
 # adds the `ion_hash` method to all simpleion value classes: